Different types of bean injection in Spring - eval - unit tests and fixes
